+++ cmd/bnscli/main.go
@@ -52,6 +52,7 @@ var commands = map[string]func(input io.Reader, output io.Writer, args []string)
 	"keyaddr":                              cmdKeyaddr,
 	"keygen":                               cmdKeygen,
 	"mnemonic":                             cmdMnemonic,
+	"msgfee-update-configuration":          cmdMsgFeeUpdateConfiguration,
 	"multisig":                             cmdMultisig,
 	"preregistration-register":             cmdPreregistrationRegister,
 	"preregistration-update-configuration": cmdPreregistrationUpdateConfiguration,